refactor(osapifix): de-main the engine into cmd/osapifix/engine
Move the rewrite engine sources, tests, surfaces, and testdata into a nested
engine package so it can be imported while still importing internal/apirev
under Go's internal rule. main.go becomes a thin shim over engine.Rewrite,
engine.Analyzers, and engine.Usage; binary behavior is unchanged.
