+++ .golangci.yml
@@ -135,6 +135,12 @@ linters:
       - linters:
           - goconst
         path: cmd/osgen/
+      # cmd/osapifix/engine tests exercise unexported engine internals
+      # (parseMajor, dirFromArg, runTypeAwareRewrite, rewriteFileTyped, ...),
+      # so they must live in package engine, not engine_test.
+      - linters:
+          - testpackage
+        path: cmd/osapifix/engine/
       - linters:
           - gosec
         path: _test\.go
