+++ cmd/osapifix/engine/hop_v3_to_v4_test.go
@@ -4,7 +4,7 @@
 // this file be licensed under the Apache-2.0 license or a
 // compatible open source license.
 
-package main
+package engine
 
 import (
 	"strings"
