+++ cmd/osapifix/engine/testdata/corpus/stub-v2/go.mod
