+++ cmd/osapifix/engine/testdata/corpus/stub-v2/opensearch.go
