+++ cmd/osapifix/engine/testdata/corpus/stub-v2/opensearchapi/opensearchapi.go
