+++ cmd/osapifix/engine/testdata/corpus/stub-v3/go.mod
