+++ cmd/osapifix/engine/testdata/corpus/stub-v3/opensearchapi/opensearchapi.go
