+++ cmd/osapifix/engine/testdata/corpus/stub-v4/opensearchapi/opensearchapi.go
