+++ cmd/osapifix/engine/testdata/corpus/v2/aliasedimport.go.golden
