+++ cmd/osapifix/engine/testdata/corpus/v2/bulk_idiom1.go
