+++ cmd/osapifix/engine/testdata/corpus/v2/carriedrootclient.go.golden
