+++ cmd/osapifix/engine/testdata/corpus/v2/paramsemit.go.golden
