+++ cmd/osapifix/engine/testdata/corpus/v2/seedops.go.golden
