+++ cmd/osapifix/engine/testdata/corpus/v3/client.go
