+++ cmd/osapifix/engine/testdata/corpus/v4/removedtype.go.golden
